+++ assets/r/dataset_load_txt.R
@@ -1,11 +1,11 @@
-# Rattle Scripts: Load a CSV file into the session as `ds`.
+# Rattle Scripts: Load a text file into the session as `txt`.
 #
 # Copyright (C) 2023, Togaware Pty Ltd.
 #
 # License: GNU General Public License, Version 3 (the "License")
 # https://www.gnu.org/licenses/gpl-3.0.en.html
 #
-# Time-stamp: <Monday 2023-10-16 09:15:33 +1100 Graham Williams>
+# Time-stamp: <Saturday 2024-11-23 10:03:01 +1100 Graham Williams>
 #
 # Licensed under the GNU General Public License, Version 3 (the "License");
 #
@@ -36,9 +36,10 @@
 #
 # https://survivor.togaware.com/datascience/csv-data-reading.html
 
+library(magrittr)        # Pipeline: %>%
 
 dsname <- "FILENAME" %>% basename() %>% sub(".txt$", "", .)
 
-ds <- readLines("FILENAME")
+txt <- readLines("FILENAME")
 
-cat(ds, sep = "\n")
+cat(txt, sep = "\n")