+++ lib/providers/keep_in_sync.dart
@@ -1,11 +1,11 @@
-/// Gloabl variable [partition].
+/// Gloabl variable [keepInSync].
 ///
-/// Copyright (C) 2023, Togaware Pty Ltd.
+/// Copyright (C) 2024, Togaware Pty Ltd.
 ///
 /// License: GNU General Public License, Version 3 (the "License")
 /// https://www.gnu.org/licenses/gpl-3.0.en.html
 //
-// Time-stamp: <Monday 2024-11-25 21:12:44 +1100 Graham Williams>
+// Time-stamp: <Sunday 2024-06-02 14:56:17 +1000 Graham Williams>
 //
 // This program is free software: you can redistribute it and/or modify it under
 // the terms of the GNU General Public License as published by the Free Software
@@ -20,12 +20,9 @@
 // You should have received a copy of the GNU General Public License along with
 // this program.  If not, see <https://www.gnu.org/licenses/>.
 ///
-/// Authors: Graham Williams
+/// Authors: Kevin Wang
 library;
 
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-// For COMP3430 2024 S2 set the default off since that is what they are asked to
-// do. Perhaps default off is good in general?
-
-final partitionProvider = StateProvider<bool>((ref) => true);
+final keepInSyncProvider = StateProvider<bool>((ref) => true);